faster distmap
git-svn-id: svn+ssh://gitlab/srv/svn-repos/pdb-redo/trunk@257 a1961a4f-ab94-4bcc-80e8-33b5a54de466

+++ include/cif++/BondMap.h
@@ -21,7 +21,7 @@ class BondMap
 	
   private:
 
-	uint32 dim;
+	size_t dim;
 	std::vector<bool> bond;
 	std::unordered_map<std::string,size_t> index;
 };
